LARGE CHANGE: Removed use of Serializable/Cloneable/Externalizable and reworked Blue's data objects to use copy constructors and DeepCopyable interface; replaced all uses of floats with doubles

+++ blue-core/src/blue/automation/ParameterTimeManagerFactory.java
@@ -34,7 +34,7 @@ public class ParameterTimeManagerFactory {
     static class DefaultParameterTimeManager implements ParameterTimeManager {
 
         @Override
-        public float getTime() {
+        public double getTime() {
             return -1.0f;
         }
