+++ blue-core/src/blue/orchestra/blueSynthBuilder/BSBVSlider.java
@@ -59,7 +59,7 @@ public class BSBVSlider extends AutomatableBSBObject implements
             if (parameters != null) {
                 Parameter param = parameters.getParameter(getObjectName());
                 if (param != null) {
-                    param.setValue((float) v);
+                    param.setValue( v);
                 }
             }
         }
@@ -156,7 +156,7 @@ public class BSBVSlider extends AutomatableBSBObject implements
 
     private static double parseNum(String string, int version) {
         if (version < 2) {
-            return (double) Float.parseFloat(string);
+            return  Double.parseDouble(string);
         }
         return Double.parseDouble(string);
     }
@@ -234,18 +234,18 @@ public class BSBVSlider extends AutomatableBSBObject implements
 //    /**
 //     * @return Returns the maximum.
 //     */
-//    public float getMaximum() {
+//    public double getMaximum() {
 //        return maximum;
 //    }
 //
 //    /**
 //     * @param maximum The maximum to set.
 //     */
-//    public void setMaximum(float maximum, boolean truncate) {
+//    public void setMaximum(double maximum, boolean truncate) {
 //        if (maximum <= getMinimum()) {
 //            return;
 //        }
-//        float oldMax = this.maximum;
+//        double oldMax = this.maximum;
 //
 //        this.maximum = maximum;
 //
@@ -269,19 +269,19 @@ public class BSBVSlider extends AutomatableBSBObject implements
 //    /**
 //     * @return Returns the minimum.
 //     */
-//    public float getMinimum() {
+//    public double getMinimum() {
 //        return minimum;
 //    }
 //
 //    /**
 //     * @param minimum The minimum to set.
 //     */
-//    public void setMinimum(float minimum, boolean truncate) {
+//    public void setMinimum(double minimum, boolean truncate) {
 //        if (minimum >= maximum) {
 //            return;
 //        }
 //
-//        float oldMin = this.minimum;
+//        double oldMin = this.minimum;
 //        this.minimum = minimum;
 //
 //        if (truncate) {
@@ -304,15 +304,15 @@ public class BSBVSlider extends AutomatableBSBObject implements
 //    /**
 //     * @return
 //     */
-//    public float getValue() {
+//    public double getValue() {
 //        return this.value;
 //    }
 //
 //    /**
 //     * @param value The value to set.
 //     */
-//    public void setValue(float value) {
-//        float oldValue = this.value;
+//    public void setValue(double value) {
+//        double oldValue = this.value;
 //        this.value = value;
 //
 //        if (getResolution() > 0) {
@@ -328,8 +328,8 @@ public class BSBVSlider extends AutomatableBSBObject implements
 //        }
 //
 //        if (propListeners != null) {
-//            propListeners.firePropertyChange("value", new Float(oldValue),
-//                    new Float(this.value));
+//            propListeners.firePropertyChange("value", new Double(oldValue),
+//                    new Double(this.value));
 //        }
 //    }
 
@@ -357,14 +357,14 @@ public class BSBVSlider extends AutomatableBSBObject implements
 //    /**
 //     * @return Returns the resolution.
 //     */
-//    public float getResolution() {
+//    public double getResolution() {
 //        return resolution;
 //    }
 //
 //    /**
 //     * @param resolution The resolution to set.
 //     */
-//    public void setResolution(float resolution) {
+//    public void setResolution(double resolution) {
 //        this.resolution = resolution;
 //
 //        if (parameters != null) {
@@ -437,32 +437,31 @@ public class BSBVSlider extends AutomatableBSBObject implements
             parameter.addParameterListener(this);
 
             if (!parameter.isAutomationEnabled()) {
-                // FIXME - make parameters use double??
-                parameter.setValue((float) getValue());
+                parameter.setValue( getValue());
             }
 
             return;
         }
 
         Parameter param = new Parameter();
-        param.setValue((float) getValue());
-        param.setMax((float) getMaximum(), true);
-        param.setMin((float) getMinimum(), true);
+        param.setValue( getValue());
+        param.setMax( getMaximum(), true);
+        param.setMin( getMinimum(), true);
         param.setName(getObjectName());
-        param.setResolution((float) getResolution());
+        param.setResolution( getResolution());
         param.addParameterListener(this);
-        param.setValue((float) getValue());
+        param.setValue( getValue());
 
         parameters.addParameter(param);
     }
 
-//    public void updateValue(float value) {
-//        float oldValue = this.value;
+//    public void updateValue(double value) {
+//        double oldValue = this.value;
 //        this.value = value;
 //
 //        if (propListeners != null) {
 //            propListeners.firePropertyChange("updateValue",
-//                    new Float(oldValue), new Float(this.value));
+//                    new Double(oldValue), new Double(this.value));
 //        }
 //    }
     @Override
@@ -470,8 +469,8 @@ public class BSBVSlider extends AutomatableBSBObject implements
         Parameter parameter = parameters.getParameter(this.objectName);
 
         if (parameter != null) {
-            float time = ParameterTimeManagerFactory.getInstance().getTime();
-            float val = parameter.getLine().getValue(time);
+            double time = ParameterTimeManagerFactory.getInstance().getTime();
+            double val = parameter.getLine().getValue(time);
 //            updateValue(val);
             setValue(val);
         }
